@@ -12,9 +12,11 @@
   "dependencies": {
     "axios": "^0.9.1",
     "express": "^4.13.4",
-    "react": "^0.14.7",
+    "react": "^0.14.8",
     "react-addons-test-utils": "^0.14.7",
     "react-dom": "^0.14.7",
+    "react-images": "^0.5.2",
+    "react-photo-gallery": "^5.0.2",
     "react-router": "^2.0.0"
   },
   "devDependencies": {
@@ -25,16 +27,22 @@
     "babel-preset-stage-0": "^6.5.0",
     "css-loader": "^0.23.1",
     "expect": "^1.14.0",
+    "fbjs": "^0.8.9",
     "foundation-sites": "^6.2.0",
-    "jquery": "^2.2.1",
+    "jquery": "^2.1.4",
     "karma": "^0.13.22",
     "karma-chrome-launcher": "^0.2.2",
     "karma-mocha": "^0.2.2",
     "karma-mocha-reporter": "^2.0.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^1.7.0",
+    "lodash": "^3.10.1",
     "mocha": "^2.4.5",
     "node-sass": "^3.4.2",
+    "object-assign": "^4.1.1",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-measure": "^1.4.5",
     "sass-loader": "^3.1.2",
     "script-loader": "^0.6.1",
     "style-loader": "^0.13.0",
