@@ -1,13 +1,18 @@
 var React = require('react');
 var Nav = require('Nav');
+var PhotoGallery = require('PhotoGallery');
+var PhotoFilterSearch = require('PhotoFilterSearch');
+
 
 var Main = (props) => {
   return (
-    <Nav/>
     <div className="row">
-      <div className="column small-centered medium-6 large-4">
+      <Nav/>
+      {/* <div className="column small-centered medium-6 large-4"> */}
+        <PhotoFilterSearch />
+        <PhotoGallery />
         {props.children}
-      </div>
+      {/* </div> */}
     </div>
   );
 };
